update styleguide-template to basic functionality

+++ packages/styleguide-template/template/tasks/version.js
@@ -8,17 +8,20 @@ const semver = require('semver');
 async function build(type) {
   const files = [
     'composer.json',
-    // 'bower.json',
-    // 'lerna.json',
+    'bower.json',
+    'lerna.json',
   ];
 
   await Promise.all(files.map(async (file) => {
     const filepath = path.join(appRootPath.toString(), file);
-    const json = JSON.parse(fs.readFileSync(filepath, 'utf8'));
-    const newVersion = semver.inc(json.version, type);
-    json.version = newVersion;
 
-    await fs.writeFileSync(filepath, JSON.stringify(json, null, 2));
+    if (fs.existsSync(filepath)) {
+      const json = JSON.parse(fs.readFileSync(filepath, 'utf8'));
+      const newVersion = semver.inc(json.version, type);
+      json.version = newVersion;
+
+      await fs.writeFileSync(filepath, JSON.stringify(json, null, 2));
+    }
   }));
 }
